Fix: User 타입 별도 파일로 관리
